I think it's waiting for WPS process to complete
