+++ requirements.txt
@@ -20,6 +20,7 @@ odc-ppt
 pylint
 pytest
 pytest-cov
+pytest-flask
 moto==1.3.16
 vega-datasets==0.9.0
 wheel
