+++ tests/conftest.py
@@ -1,0 +1,7 @@
+import pytest
+from datacube_wps import create_app
+
+
+@pytest.fixture
+def app():
+    return create_app()
